+++ mathesar_ui/src/pages/data-explorer/DataExplorerPage.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import type { SchemaEntry } from '@mathesar/AppTypes';
   import LayoutWithHeader from '@mathesar/layouts/LayoutWithHeader.svelte';
-  import { DataExplorer } from '@mathesar/systems/query-builder';
-  import type { QueryManager } from '@mathesar/systems/query-builder/types';
+  import { DataExplorer } from '@mathesar/systems/data-explorer';
+  import type { QueryManager } from '@mathesar/systems/data-explorer/types';
 
   export let schema: SchemaEntry;
   export let queryManager: QueryManager;
