+++ mathesar_ui/src/systems/data-explorer/QuerySummarizationTransformationModel.ts
