+++ mathesar_ui/src/systems/data-explorer/QueryUndoRedoManager.ts
