+++ mathesar_ui/src/systems/data-explorer/column-selection-pane/TableGroupCollapsible.svelte
