+++ mathesar_ui/src/systems/data-explorer/output-config-sidebar/TransformationsPane.svelte
