+++ mathesar_ui/src/systems/data-explorer/output-config-sidebar/summarization/Aggregation.svelte
