+++ mathesar_ui/src/systems/data-explorer/result-pane/ResultPane.svelte
